Remove BaseMessageTypes.h and move its contents into MessageTypes.h
