+++ dds/DCPS/RTPS/BaseMessageUtils.cpp
@@ -4,7 +4,7 @@
  */
 
 #include "BaseMessageUtils.h"
-
+#include "MessageTypes.h"
 #include <dds/DCPS/Time_Helper.h>
 #include <dds/OpenddsDcpsExtTypeSupportImpl.h>
