+++ dds/DCPS/transport/rtps_udp/RtpsUdpReceiveStrategy.cpp
@@ -10,7 +10,6 @@
 #include "RtpsUdpInst.h"
 #include "RtpsUdpTransport.h"
 
-#include "dds/DCPS/RTPS/BaseMessageTypes.h"
 #include "dds/DCPS/RTPS/BaseMessageUtils.h"
 #include "dds/DCPS/RTPS/MessageTypes.h"
 #include "dds/DCPS/RTPS/Logging.h"
